tienda-app/jacoyitos-backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
flask-bcrypt
pyjwt
